Move tests to root_test.go
